+++ src/nsidc/metgen/config.py
@@ -59,19 +59,19 @@ class Config:
     def show(self):
         # TODO: add section headings in the right spot
         #       (if we think we need them in the output)
-        LOGGER = logging.getLogger(constants.ROOT_LOGGER)
-        LOGGER.info("")
-        LOGGER.info("Using configuration:")
+        logger = logging.getLogger(constants.ROOT_LOGGER)
+        logger.info_plus("")
+        logger.info_plus("Using configuration:")
         for k, v in self.__dict__.items():
-            LOGGER.info(f"  + {k}: {v}")
+            logger.info_plus(f"  + {k}: {v}")
 
         if self.dry_run:
-            LOGGER.info("")
-            LOGGER.info(
+            logger.info_plus("")
+            logger.info_plus(
                 "Note: The dry-run option was included, so no files will be \
 staged and no CNM messages published."
             )
-            LOGGER.info("")
+            logger.info_plus("")
 
     def ummg_path(self):
         return Path(self.local_output_dir, self.ummg_dir)